pyproject.toml
[tool.poetry]
name = "aviahome"
version = "0.1.0"
description = ""
authors = ["JEthebest <iskokairatov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
pyaviation = "^0.0.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
